+++ packages/menu-dropdown/src/MenuDropdownContent/index.tsx
@@ -3,11 +3,11 @@ import { MenuDropdownItem, ItemProps } from '../MenuDropdownItem';
 import { ContentWrapper } from './index.style';
 import { MenuDropdownDivider } from '../MenuDropdownDivider';
 
-export class MenuDropdownContent extends React.Component<{}, {}> {
+export class MenuDropdownContent extends React.Component {
   static Item = MenuDropdownItem;
   static Divider = MenuDropdownDivider;
 
-  private renderItem = (child: React.ReactChild): React.ReactNode => {
+  private renderItem = (child: React.ReactChild): React.ReactChild | undefined => {
     if (React.isValidElement<ItemProps>(child)) {
       return child;
     }
@@ -16,6 +16,6 @@ export class MenuDropdownContent extends React.Component<{}, {}> {
 
   render() {
     const { children } = this.props;
-    return <ContentWrapper>{React.Children.map(children, this.renderItem)}</ContentWrapper>;
+    return <ContentWrapper>{React.Children.map(children as React.ReactChild, this.renderItem)}</ContentWrapper>;
   }
 }