+++ packages/tabs/src/TabContext.tsx
@@ -1,4 +1,6 @@
 import * as createReactContext from 'create-react-context';
+// @ts-ignore
+const PonyfillContext = typeof createReactContext === 'object' ? createReactContext.default : createReactContext;
 
 export interface TabContextState {
   tabCount: number;
@@ -9,6 +11,6 @@ export interface TabContextState {
 // Supressing "Cannot invoke an expression whose type lacks a call signature." error
 // More details here: https://github.com/jamiebuilds/create-react-context/pull/20
 // @ts-ignore
-const TabContext = createReactContext<TabContextState>({});
+const TabContext = PonyfillContext<TabContextState>({});
 
 export default TabContext;