+++ packages/utils/src/index.ts
@@ -37,7 +37,7 @@ interface DataObject {
  * return all the props of an object where the key begins with 'data-'
  * @param props The props object of a given component
  */
-export const getDataProps = (props: ComponentProps) => {
+export const getDataProps = (props: ComponentProps): DataObject => {
   const dataProps: DataObject = {};
   for (const key in props) {
     if (key.startsWith('data-')) {