+++ website/src/modules/Package/index.tsx
@@ -1,22 +1,16 @@
 import * as React from 'react';
 import { Switch, Route } from 'react-router-dom';
 import metadata, { Workspace } from '../../utils/info';
+import { RouteMatch } from '../../utils/DocumentViewer';
 import { Layout } from '../Layout';
 import { Header } from './Header';
 import { Overview } from './Overview';
 import { Doc } from './Doc';
 import { Content } from './index.style';
 
-export interface PackageProps {
-  match: {
-    path: string;
-    params: {
-      pkg: string;
-    };
-  };
-}
+export type PackageExampleProps = {} & RouteMatch;
 
-export class Package extends React.Component<PackageProps> {
+export class Package extends React.Component<PackageExampleProps> {
   get workspace(): Workspace {
     const {
       match: {
@@ -35,10 +29,10 @@ export class Package extends React.Component<PackageProps> {
         <Content>
           <Switch>
             <Route path={`${match.path}`} exact>
-              {(props) => <Overview {...props} workspace={this.workspace} />}
+              {(props: RouteMatch) => <Overview {...props} workspace={this.workspace} />}
             </Route>
             <Route path={`${match.path}/:doc`} exact>
-              {(props) => <Doc {...props} workspace={this.workspace} />}
+              {(props: RouteMatch) => <Doc {...props} workspace={this.workspace} />}
             </Route>
           </Switch>
         </Content>